moved setModel and getModel methods
