+++ ContextDesktop/src/main/java/com/github/migi_1/Context/model/entity/Platform.java
@@ -2,7 +2,6 @@ package com.github.migi_1.Context.model.entity;
 
 import com.github.migi_1.Context.utility.ProjectAssetManager;
 import com.jme3.math.Vector3f;
-import com.jme3.scene.Spatial;
 
 /**
  * Class that handles everything that have to do with the platform.
@@ -15,7 +14,6 @@ import com.jme3.scene.Spatial;
 public class Platform extends Entity {
 
     private static final String PATHNAME = "Models/testPlatform.j3o";
-    private Spatial model;
 
 
     /**
@@ -23,27 +21,9 @@ public class Platform extends Entity {
      * @param startLocation location where the carrier will be initialized
      */
     public Platform(Vector3f startLocation) {
-        model = ProjectAssetManager.getInstance().getAssetManager().loadModel(PATHNAME);
-        model.setLocalTranslation(startLocation);
+        setModel(ProjectAssetManager.getInstance().getAssetManager().loadModel(PATHNAME));
+        getModel().setLocalTranslation(startLocation);
         setMoveBehaviour(new AcceleratorMoveBehaviour());
-    }
-
-    /**
-     * Returns the model of the platform.
-     * @return
-     */
-    @Override
-    public Spatial getModel() {
-        return model;
-    }
-
-    /**
-     * Sets the model of the platform.
-     * @param model
-     */
-    @Override
-    public void setModel(Spatial model) {
-        this.model = model;
     }
